+++ src/Account.h
@@ -25,6 +25,8 @@
 #include "Authentication.h"
 #include "SwiftResult.h"
 #include "Header.h"
+
+#include <cstdint>
 #include <iostream>
 #include <vector>
 
@@ -91,7 +93,7 @@ private:
   /**
    * Number of calls made to Object Store
    */
-  static ulong numOfCalls;
+  static uint32_t numOfCalls;
 
 public:
   virtual ~Account();
@@ -138,14 +140,14 @@ public:
   /**
    * Increase the call counter which tracks how many calls are made to the ObjectStore
    */
-  static ulong increaseCallCounter();
+  static uint32_t increaseCallCounter();
 
   /**
    * Returns the number of HTTP calls that have been made to the ObjectStore server. This could be
    * useful to check the efficiency of the methods and configuration in use
    * @return number of calls made to the ObjectStore server
    */
-  static ulong getNumberOfCalls();
+  static uint32_t getNumberOfCalls();
 
   /**
    * Returns the tenant of the account.
